requirements.txt
accelerate==1.1.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
attrs==24.2.0
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
datasets==3.1.0
dill==0.3.8
eval_type_backport==0.2.0
filelock==3.16.1
Flask==3.0.3
frozenlist==1.5.0
fsspec==2024.9.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
huggingface-hub==0.26.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
jsonpath-python==1.0.6
llmlingua==0.2.2
MarkupSafe==3.0.2
mistralai==1.1.0
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
mypy-extensions==1.0.0
networkx==3.4.2
nltk==3.9.1
numpy==2.1.3
packaging==24.1
pandas==2.2.3
propcache==0.2.0
psutil==6.1.0
pyarrow==18.0.0
pydantic==2.9.2
pydantic_core==2.23.4
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
regex==2024.9.11
requests==2.32.3
safetensors==0.4.5
setuptools==75.3.0
six==1.16.0
sniffio==1.3.1
sympy==1.13.1
tiktoken==0.8.0
tokenizers==0.20.1
torch==2.5.1
tqdm==4.66.6
transformers==4.46.1
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
Werkzeug==3.1.1
xxhash==3.5.0
yarl==1.17.1
